+++ threedigrid_builder/grid/grid.py
@@ -91,7 +91,7 @@ class Grid:
         lines: Lines,
         pumps: Optional[Pumps] = None,
         cross_sections: Optional[CrossSections] = None,
-        embedded_nodes=None,
+        nodes_embedded=None,
         meta=None,
         quadtree_stats=None,
     ):
@@ -109,17 +109,17 @@ class Grid:
             raise TypeError(
                 f"Expected CrossSections instance, got {type(cross_sections)}"
             )
-        if embedded_nodes is None:
-            embedded_nodes = Nodes(id=[])
-        elif not isinstance(embedded_nodes, Nodes):
-            raise TypeError(f"Expected Nodes instance, got {type(embedded_nodes)}")
+        if nodes_embedded is None:
+            nodes_embedded = Nodes(id=[])
+        elif not isinstance(nodes_embedded, Nodes):
+            raise TypeError(f"Expected Nodes instance, got {type(nodes_embedded)}")
         self.nodes = nodes
         self.lines = lines
         self.meta = meta
         self.quadtree_stats = quadtree_stats
         self.pumps = pumps
         self.cross_sections = cross_sections
-        self.embedded_nodes = embedded_nodes
+        self.nodes_embedded = nodes_embedded
         self._cell_tree = None
 
     def __add__(self, other):
@@ -130,7 +130,7 @@ class Grid:
                 "equal types."
             )
         new_attrs = {}
-        for name in ("nodes", "lines", "embedded_nodes"):
+        for name in ("nodes", "lines", "nodes_embedded"):
             new_attrs[name] = getattr(self, name) + getattr(other, name)
         for name in ("meta", "quadtree_stats", "pumps", "cross_sections"):
             if getattr(other, name) is None:
@@ -536,9 +536,13 @@ class Grid:
     def embed_nodes(self):
         """Integrate embedded connection nodes into the 2D cells.
 
-        grid.embedded_nodes will contain the removed (embedded) nodes.
+        grid.nodes_embedded will contain the removed (embedded) nodes.
         """
-        self.embedded_nodes = embed_nodes(self)
+        # before connecting lines to 2D nodes, first set the line_coords based on the
+        # original coordinate of the embedded nodes:
+        self.lines.set_line_coords(self.nodes)
+
+        self.nodes_embedded = embed_nodes(self)
 
     def add_1d2d(
         self, connection_nodes, channels, pipes, locations, culverts, line_id_counter